+++ source/scan-tool/ChangeLog.md
@@ -1,5 +1,11 @@
 # ChangeLog for scan-tool
 
+## Version 0.26 (2016-01-09)
+  - change directory structure of request cache (again)
+  - add option for automatic transition from older request cache structures
+    (both the first from v0.20 till v0.21 and the later structure from v0.22
+    till v0.25) to new request cache structure
+
 ## Version 0.25 (2016-01-04)
   - improve rescan requests